create export/ package; convert deploy/ to deprecation shim
- Add src/rfdetr/export/ as the canonical ONNX export package (copy of deploy/)
- Update all internal rfdetr.deploy.* imports to rfdetr.export.* inside export/
- Convert src/rfdetr/deploy/ (__init__.py, export.py, _onnx/__init__.py) to
  DeprecationWarning re-export shims pointing to rfdetr.export.*
- Update detr.py lazy import: rfdetr.deploy.export → rfdetr.export.export
- Update tests/models/test_export.py: import canonical rfdetr.export.export module
  (patch.object targets must match the module where the function is defined)
