+++ src/rfdetr/deploy/_onnx/__init__.py
@@ -3,15 +3,16 @@
 # Copyright (c) 2025 Roboflow. All Rights Reserved.
 # Licensed under the Apache License, Version 2.0 [see LICENSE for details]
 # ------------------------------------------------------------------------
-# Copied and modified from LW-DETR (https://github.com/Atten4Vis/LW-DETR)
-# Copyright (c) 2024 Baidu. All Rights Reserved.
-# Licensed under the Apache License, Version 2.0 [see LICENSE for details]
-# ------------------------------------------------------------------------
 
-"""
-onnx optimizer and symbolic registry
-"""
+"""Backward-compatibility shim — rfdetr.deploy._onnx is deprecated; use rfdetr.export._onnx."""
 
-from rfdetr.deploy._onnx import optimizer, symbolic
-from rfdetr.deploy._onnx.optimizer import OnnxOptimizer
-from rfdetr.deploy._onnx.symbolic import CustomOpSymbolicRegistry
+import warnings
+
+warnings.warn(
+    "rfdetr.deploy._onnx is deprecated; use rfdetr.export._onnx instead.",
+    DeprecationWarning,
+    stacklevel=2,
+)
+
+from rfdetr.export._onnx import *  # noqa: F401, F403, E402
+from rfdetr.export._onnx import OnnxOptimizer, CustomOpSymbolicRegistry  # noqa: E402
