+++ src/rfdetr/deploy/benchmark.py
@@ -3,542 +3,16 @@
 # Copyright (c) 2025 Roboflow. All Rights Reserved.
 # Licensed under the Apache License, Version 2.0 [see LICENSE for details]
 # ------------------------------------------------------------------------
-# Copied and modified from LW-DETR (https://github.com/Atten4Vis/LW-DETR)
-# Copyright (c) 2024 Baidu. All Rights Reserved.
-# ------------------------------------------------------------------------
 
-"""
-This tool provides performance benchmarks by using ONNX Runtime and TensorRT
-to run inference on a given model with the COCO validation set. It offers
-reliable measurements of inference latency using ONNX Runtime or TensorRT
-on the device.
-"""
+"""Backward-compatibility shim — rfdetr.deploy.benchmark is deprecated; use rfdetr.export.benchmark."""
 
-import argparse
-import contextlib
-import copy
-import json
-import os
-import os.path as osp
-import random
-import time
-from collections import OrderedDict, namedtuple
+import warnings
 
-import numpy as np
-import supervision as sv
-import torch
-import torchvision.transforms.functional as F
-from PIL import Image
-from pycocotools.coco import COCO
-from pycocotools.cocoeval import COCOeval
-from tqdm.auto import tqdm
+warnings.warn(
+    "rfdetr.deploy.benchmark is deprecated; use rfdetr.export.benchmark instead.",
+    DeprecationWarning,
+    stacklevel=2,
+)
 
-try:
-    import tensorrt as trt
-except ImportError:
-    trt = None
-
-try:
-    import pycuda.driver as cuda
-except ImportError:
-    cuda = None
-
-from rfdetr.utilities.box_ops import box_xyxy_to_cxcywh
-from rfdetr.utilities.logger import get_logger
-
-logger = get_logger()
-
-
-def parser_args():
-    parser = argparse.ArgumentParser("performance benchmark tool for onnx/trt model")
-    parser.add_argument("--path", type=str, help="engine file path")
-    parser.add_argument("--coco_path", type=str, default="data/coco", help="coco dataset path")
-    parser.add_argument("--device", default=0, type=int)
-    parser.add_argument("--run_benchmark", action="store_true", help="repeat the inference to benchmark the latency")
-    parser.add_argument("--disable_eval", action="store_true", help="disable evaluation")
-    return parser.parse_args()
-
-
-class CocoEvaluator(object):
-    def __init__(self, coco_gt, iou_types):
-        assert isinstance(iou_types, (list, tuple))
-        coco_gt = COCO(coco_gt)
-        coco_gt = copy.deepcopy(coco_gt)
-        self.coco_gt = coco_gt
-
-        self.iou_types = iou_types
-        self.coco_eval = {}
-        for iou_type in iou_types:
-            self.coco_eval[iou_type] = COCOeval(coco_gt, iouType=iou_type)
-
-        self.img_ids = []
-        self.eval_imgs = {k: [] for k in iou_types}
-
-    def update(self, predictions):
-        img_ids = list(np.unique(list(predictions.keys())))
-        self.img_ids.extend(img_ids)
-
-        for iou_type in self.iou_types:
-            results = self.prepare(predictions, iou_type)
-
-            # suppress pycocotools prints
-            with open(os.devnull, "w") as devnull:
-                with contextlib.redirect_stdout(devnull):
-                    coco_dt = COCO.loadRes(self.coco_gt, results) if results else COCO()
-            coco_eval = self.coco_eval[iou_type]
-
-            coco_eval.cocoDt = coco_dt
-            coco_eval.params.imgIds = list(img_ids)
-            img_ids, eval_imgs = evaluate(coco_eval)
-
-            self.eval_imgs[iou_type].append(eval_imgs)
-
-    def synchronize_between_processes(self):
-        for iou_type in self.iou_types:
-            self.eval_imgs[iou_type] = np.concatenate(self.eval_imgs[iou_type], 2)
-            create_common_coco_eval(self.coco_eval[iou_type], self.img_ids, self.eval_imgs[iou_type])
-
-    def accumulate(self):
-        for coco_eval in self.coco_eval.values():
-            coco_eval.accumulate()
-
-    def summarize(self):
-        for iou_type, coco_eval in self.coco_eval.items():
-            logger.info("IoU metric: {}".format(iou_type))
-            coco_eval.summarize()
-
-    def prepare(self, predictions, iou_type):
-        if iou_type == "bbox":
-            return self.prepare_for_coco_detection(predictions)
-        else:
-            raise ValueError("Unknown iou type {}".format(iou_type))
-
-    def prepare_for_coco_detection(self, predictions):
-        coco_results = []
-        for original_id, prediction in predictions.items():
-            if len(prediction) == 0:
-                continue
-
-            boxes = prediction["boxes"]
-            boxes = sv.xyxy_to_xywh(boxes.cpu().numpy()).tolist()
-            scores = prediction["scores"].tolist()
-            labels = prediction["labels"].tolist()
-
-            coco_results.extend(
-                [
-                    {
-                        "image_id": original_id,
-                        "category_id": labels[k],
-                        "bbox": box,
-                        "score": scores[k],
-                    }
-                    for k, box in enumerate(boxes)
-                ]
-            )
-        return coco_results
-
-
-def create_common_coco_eval(coco_eval, img_ids, eval_imgs):
-    img_ids = list(img_ids)
-    eval_imgs = list(eval_imgs.flatten())
-
-    coco_eval.evalImgs = eval_imgs
-    coco_eval.params.imgIds = img_ids
-    coco_eval._paramsEval = copy.deepcopy(coco_eval.params)
-
-
-def evaluate(self):
-    """
-    Run per image evaluation on given images and store results (a list of dict) in self.evalImgs
-    :return: None
-    """
-    # Running per image evaluation...
-    p = self.params
-    # add backward compatibility if useSegm is specified in params
-    if p.useSegm is not None:
-        p.iouType = "segm" if p.useSegm == 1 else "bbox"
-        logger.warning("useSegm (deprecated) is not None. Running {} evaluation".format(p.iouType))
-    # print('Evaluate annotation type *{}*'.format(p.iouType))
-    p.imgIds = list(np.unique(p.imgIds))
-    if p.useCats:
-        p.catIds = list(np.unique(p.catIds))
-    p.maxDets = sorted(p.maxDets)
-    self.params = p
-
-    self._prepare()
-    # loop through images, area range, max detection number
-    catIds = p.catIds if p.useCats else [-1]
-
-    if p.iouType == "segm" or p.iouType == "bbox":
-        computeIoU = self.computeIoU
-    elif p.iouType == "keypoints":
-        computeIoU = self.computeOks
-    self.ious = {(imgId, catId): computeIoU(imgId, catId) for imgId in p.imgIds for catId in catIds}
-
-    evaluateImg = self.evaluateImg
-    maxDet = p.maxDets[-1]
-    evalImgs = [
-        evaluateImg(imgId, catId, areaRng, maxDet) for catId in catIds for areaRng in p.areaRng for imgId in p.imgIds
-    ]
-    # this is NOT in the pycocotools code, but could be done outside
-    evalImgs = np.asarray(evalImgs).reshape(len(catIds), len(p.areaRng), len(p.imgIds))
-    self._paramsEval = copy.deepcopy(self.params)
-    return p.imgIds, evalImgs
-
-
-def get_image_list(ann_file):
-    with open(ann_file, "r") as fin:
-        data = json.load(fin)
-    return data["images"]
-
-
-def load_image(file_path):
-    return Image.open(file_path).convert("RGB")
-
-
-def infer_transforms():
-    from torchvision.transforms.v2 import Compose, Resize, ToDtype, ToImage
-
-    from rfdetr.datasets.transforms import Normalize
-
-    return Compose(
-        [
-            Resize((640, 640)),
-            ToImage(),
-            ToDtype(torch.float32, scale=True),
-            Normalize(),
-        ]
-    )
-
-
-def box_cxcywh_to_xyxy(x):
-    x_c, y_c, w, h = x.unbind(-1)
-    b = [
-        (x_c - 0.5 * w.clamp(min=0.0)),
-        (y_c - 0.5 * h.clamp(min=0.0)),
-        (x_c + 0.5 * w.clamp(min=0.0)),
-        (y_c + 0.5 * h.clamp(min=0.0)),
-    ]
-    return torch.stack(b, dim=-1)
-
-
-def post_process(outputs, target_sizes):
-    out_logits, out_bbox = outputs["labels"], outputs["dets"]
-
-    assert len(out_logits) == len(target_sizes)
-    assert target_sizes.shape[1] == 2
-
-    prob = out_logits.sigmoid()
-    topk_values, topk_indexes = torch.topk(prob.view(out_logits.shape[0], -1), 300, dim=1)
-    scores = topk_values
-    topk_boxes = topk_indexes // out_logits.shape[2]
-    labels = topk_indexes % out_logits.shape[2]
-    boxes = box_cxcywh_to_xyxy(out_bbox)
-    boxes = torch.gather(boxes, 1, topk_boxes.unsqueeze(-1).repeat(1, 1, 4))
-
-    # and from relative [0, 1] to absolute [0, height] coordinates
-    img_h, img_w = target_sizes.unbind(1)
-    scale_fct = torch.stack([img_w, img_h, img_w, img_h], dim=1)
-    boxes = boxes * scale_fct[:, None, :]
-
-    results = [{"scores": s, "labels": l, "boxes": b} for s, l, b in zip(scores, labels, boxes)]
-
-    return results
-
-
-def infer_onnx(sess, coco_evaluator, time_profile, prefix, img_list, device, repeats=1):
-    time_list = []
-    for img_dict in tqdm(img_list):
-        image = load_image(os.path.join(prefix, img_dict["file_name"]))
-        width, height = image.size
-        orig_target_sizes = torch.Tensor([height, width])
-        image_tensor, _ = infer_transforms()(image, None)  # target is None
-
-        samples = image_tensor[None].numpy()
-
-        time_profile.reset()
-        with time_profile:
-            for _ in range(repeats):
-                res = sess.run(None, {"input": samples})
-        time_list.append(time_profile.total / repeats)
-        outputs = {}
-        outputs["labels"] = torch.Tensor(res[1]).to(device)
-        outputs["dets"] = torch.Tensor(res[0]).to(device)
-
-        orig_target_sizes = torch.stack([orig_target_sizes], dim=0).to(device)
-        results = post_process(outputs, orig_target_sizes)
-        res = {img_dict["id"]: results[0]}
-        if coco_evaluator is not None:
-            coco_evaluator.update(res)
-
-    logger.info("Model latency with ONNX Runtime: {}ms".format(1000 * sum(time_list) / len(img_list)))
-
-    # accumulate predictions from all images
-    stats = {}
-    if coco_evaluator is not None:
-        coco_evaluator.synchronize_between_processes()
-        coco_evaluator.accumulate()
-        coco_evaluator.summarize()
-        stats["coco_eval_bbox"] = coco_evaluator.coco_eval["bbox"].stats.tolist()
-        logger.info(stats)
-
-
-def infer_engine(model, coco_evaluator, time_profile, prefix, img_list, device, repeats=1):
-    time_list = []
-    for img_dict in tqdm(img_list):
-        image = load_image(os.path.join(prefix, img_dict["file_name"]))
-        width, height = image.size
-        orig_target_sizes = torch.Tensor([height, width])
-        image_tensor, _ = infer_transforms()(image, None)  # target is None
-
-        samples = image_tensor[None].to(device)
-        _, _, h, w = samples.shape
-        # torch.Tensor(np.array([h, w]).reshape((1, 2)).astype(np.float32)).to(device)
-        # torch.Tensor(np.array([h / height, w / width]).reshape((1, 2)).astype(np.float32)).to(device)
-
-        time_profile.reset()
-        with time_profile:
-            for _ in range(repeats):
-                outputs = model({"input": samples})
-
-        time_list.append(time_profile.total / repeats)
-        orig_target_sizes = torch.stack([orig_target_sizes], dim=0).to(device)
-        if coco_evaluator is not None:
-            results = post_process(outputs, orig_target_sizes)
-            res = {img_dict["id"]: results[0]}
-            coco_evaluator.update(res)
-
-    logger.info("Model latency with TensorRT: {}ms".format(1000 * sum(time_list) / len(img_list)))
-
-    # accumulate predictions from all images
-    stats = {}
-    if coco_evaluator is not None:
-        coco_evaluator.synchronize_between_processes()
-        coco_evaluator.accumulate()
-        coco_evaluator.summarize()
-        stats["coco_eval_bbox"] = coco_evaluator.coco_eval["bbox"].stats.tolist()
-        logger.info(stats)
-
-
-class TRTInference(object):
-    """TensorRT inference engine"""
-
-    def __init__(
-        self, engine_path="dino.engine", device="cuda:0", sync_mode: bool = False, max_batch_size=32, verbose=False
-    ):
-        if not trt:
-            raise ImportError("TensorRT is not installed. Please install TensorRT to use TRTInference.")
-
-        self.engine_path = engine_path
-        self.device = device
-        self.sync_mode = sync_mode
-        self.max_batch_size = max_batch_size
-
-        self.logger = trt.Logger(trt.Logger.VERBOSE) if verbose else trt.Logger(trt.Logger.INFO)
-
-        self.engine = self.load_engine(engine_path)
-
-        self.context = self.engine.create_execution_context()
-
-        self.bindings = self.get_bindings(self.engine, self.context, self.max_batch_size, self.device)
-        self.bindings_addr = OrderedDict((n, v.ptr) for n, v in self.bindings.items())
-
-        self.input_names = self.get_input_names()
-        self.output_names = self.get_output_names()
-        self.stream = None
-
-        if not self.sync_mode:
-            if not cuda:
-                raise ImportError(
-                    "pycuda is not installed. Please install `pycuda` to use TRTInference with async mode."
-                )
-
-            self.stream = cuda.Stream()
-
-        # self.time_profile = TimeProfiler()
-        self.time_profile = None
-
-    def get_dummy_input(self, batch_size: int):
-        blob = {}
-        for name, binding in self.bindings.items():
-            if self.engine.get_tensor_mode(name) == trt.TensorIOMode.INPUT:
-                logger.info(f"make dummy input {name} with shape {binding.shape}")
-                blob[name] = torch.rand(batch_size, *binding.shape[1:]).float().to("cuda:0")
-        return blob
-
-    def load_engine(self, path):
-        """load engine"""
-        trt.init_libnvinfer_plugins(self.logger, "")
-        with open(path, "rb") as f, trt.Runtime(self.logger) as runtime:
-            return runtime.deserialize_cuda_engine(f.read())
-
-    def get_input_names(
-        self,
-    ):
-        names = []
-        for _, name in enumerate(self.engine):
-            if self.engine.get_tensor_mode(name) == trt.TensorIOMode.INPUT:
-                names.append(name)
-        return names
-
-    def get_output_names(
-        self,
-    ):
-        names = []
-        for _, name in enumerate(self.engine):
-            if self.engine.get_tensor_mode(name) == trt.TensorIOMode.OUTPUT:
-                names.append(name)
-        return names
-
-    def get_bindings(self, engine, context, max_batch_size=32, device=None):
-        """build binddings"""
-        Binding = namedtuple("Binding", ("name", "dtype", "shape", "data", "ptr"))
-        bindings = OrderedDict()
-
-        for i, name in enumerate(engine):
-            shape = engine.get_tensor_shape(name)
-            dtype = trt.nptype(engine.get_tensor_dtype(name))
-
-            if shape[0] == -1:
-                raise NotImplementedError
-
-            else:
-                data = torch.from_numpy(np.empty(shape, dtype=dtype)).to(device)
-                bindings[name] = Binding(name, dtype, shape, data, data.data_ptr())
-
-        return bindings
-
-    def run_sync(self, blob):
-        self.bindings_addr.update({n: blob[n].data_ptr() for n in self.input_names})
-        self.context.execute_v2(list(self.bindings_addr.values()))
-        outputs = {n: self.bindings[n].data for n in self.output_names}
-        return outputs
-
-    def run_async(self, blob):
-        self.bindings_addr.update({n: blob[n].data_ptr() for n in self.input_names})
-        bindings_addr = [int(v) for _, v in self.bindings_addr.items()]
-        self.context.execute_async_v2(bindings=bindings_addr, stream_handle=self.stream.handle)
-        outputs = {n: self.bindings[n].data for n in self.output_names}
-        self.stream.synchronize()
-        return outputs
-
-    def __call__(self, blob):
-        if self.sync_mode:
-            return self.run_sync(blob)
-        else:
-            return self.run_async(blob)
-
-    def synchronize(
-        self,
-    ):
-        if self.sync_mode:
-            if torch.cuda.is_available():
-                torch.cuda.synchronize()
-            return
-
-        if self.stream is not None:
-            self.stream.synchronize()
-        elif torch.cuda.is_available():
-            torch.cuda.synchronize()
-
-    def speed(self, blob, n):
-        self.time_profile.reset()
-        with self.time_profile:
-            for _ in range(n):
-                _ = self(blob)
-        return self.time_profile.total / n
-
-    def build_engine(self, onnx_file_path, engine_file_path, max_batch_size=32):
-        """Takes an ONNX file and creates a TensorRT engine to run inference with
-        http://gitlab.baidu.com/paddle-inference/benchmark/blob/main/backend_trt.py#L57
-        """
-        EXPLICIT_BATCH = 1 << (int)(trt.NetworkDefinitionCreationFlag.EXPLICIT_BATCH)
-        with (
-            trt.Builder(self.logger) as builder,
-            builder.create_network(EXPLICIT_BATCH) as network,
-            trt.OnnxParser(network, self.logger) as parser,
-            builder.create_builder_config() as config,
-        ):
-            config.set_memory_pool_limit(trt.MemoryPoolType.WORKSPACE, 1 << 30)  # 1024 MiB
-            config.set_flag(trt.BuilderFlag.FP16)
-
-            with open(onnx_file_path, "rb") as model:
-                if not parser.parse(model.read()):
-                    logger.error("ERROR: Failed to parse the ONNX file.")
-                    for error in range(parser.num_errors):
-                        logger.error(parser.get_error(error))
-                    return None
-
-            serialized_engine = builder.build_serialized_network(network, config)
-            with open(engine_file_path, "wb") as f:
-                f.write(serialized_engine)
-
-            return serialized_engine
-
-
-class TimeProfiler(contextlib.ContextDecorator):
-    def __init__(
-        self,
-    ):
-        self.total = 0
-
-    def __enter__(
-        self,
-    ):
-        self.start = self.time()
-        return self
-
-    def __exit__(self, type, value, traceback):
-        self.total += self.time() - self.start
-
-    def reset(
-        self,
-    ):
-        self.total = 0
-
-    def time(
-        self,
-    ):
-        if torch.cuda.is_available():
-            torch.cuda.synchronize()
-        return time.perf_counter()
-
-
-def main(args):
-    logger.info(args)
-
-    coco_gt = osp.join(args.coco_path, "annotations/instances_val2017.json")
-    img_list = get_image_list(coco_gt)
-    prefix = osp.join(args.coco_path, "val2017")
-    if args.run_benchmark:
-        repeats = 10
-        logger.info(
-            "Inference for each image will be repeated 10 times to obtain a reliable measurement of inference latency."
-        )
-    else:
-        repeats = 1
-
-    if args.disable_eval:
-        coco_evaluator = None
-    else:
-        coco_evaluator = CocoEvaluator(coco_gt, ("bbox",))
-
-    time_profile = TimeProfiler()
-
-    if args.path.endswith(".onnx"):
-        import onnxruntime as nxrun
-
-        sess = nxrun.InferenceSession(args.path, providers=["CUDAExecutionProvider"])
-        infer_onnx(sess, coco_evaluator, time_profile, prefix, img_list, device=f"cuda:{args.device}", repeats=repeats)
-    elif args.path.endswith(".engine"):
-        model = TRTInference(args.path, sync_mode=True, device=f"cuda:{args.device}")
-        infer_engine(
-            model, coco_evaluator, time_profile, prefix, img_list, device=f"cuda:{args.device}", repeats=repeats
-        )
-    else:
-        raise NotImplementedError('Only model file names ending with ".onnx" and ".engine" are supported.')
-
-
-if __name__ == "__main__":
-    args = parser_args()
-    main(args)
+from rfdetr.export.benchmark import *  # noqa: F401, F403, E402
+from rfdetr.export.benchmark import TRTInference, infer_transforms  # noqa: E402
