+++ src/rfdetr/export/__init__.py
@@ -3,15 +3,3 @@
 # Copyright (c) 2025 Roboflow. All Rights Reserved.
 # Licensed under the Apache License, Version 2.0 [see LICENSE for details]
 # ------------------------------------------------------------------------
-
-"""Backward-compatibility shim — rfdetr.deploy is deprecated; use rfdetr.export."""
-
-import warnings
-
-warnings.warn(
-    "rfdetr.deploy is deprecated; use rfdetr.export instead.",
-    DeprecationWarning,
-    stacklevel=2,
-)
-
-from rfdetr.export import *  # noqa: F401, F403, E402
